# todo-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/todos": {
            "get": {
                "description": "get all ToDos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "get ToDOs",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ToDo"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "create new ToDo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "add ToDo",
                "parameters": [
                    {
                        "description": "ToDo details",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                }
            }
        },
        "/todos/{id}": {
            "get": {
                "description": "Get existing ToDo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Get ToDo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ToDo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                }
            },
            "put": {
                "description": "Update existing ToDo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Update ToDo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ToDo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ToDo details",
                        "name": "todo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete existing ToDo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo"
                ],
                "summary": "Delete ToDo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ToDo ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ToDo": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2023-09-12T15:45:44.523792+03:00"
                },
                "description": {
                    "type": "string",
                    "example": "My New ToDo"
                },
                "id": {
                    "type": "string",
                    "example": "2afbae48-4fa5-41f6-9a2f-82b9503d18ed"
                },
                "name": {
                    "type": "string",
                    "example": "ToDo14"
                },
                "status": {
                    "type": "string",
                    "example": "New"
                },
                "updatedAt": {
                    "type": "string",
                    "example": "2023-09-12T15:45:44.523792+03:00"
                }
            }
        }
    }
}